Language change option

+++ lhc_web/design/defaulttheme/tpl/pagelayouts/parts/page_head_js.tpl.php
@@ -5,9 +5,9 @@ var WWW_DIR_JAVASCRIPT_FILES_NOTIFICATION = '<?php echo erLhcoreClassDesign::des
 var confLH = {};
 confLH.back_office_sinterval = <?php echo (int)(erConfigClassLhConfig::getInstance()->getSetting('chat','back_office_sinterval')*1000) ?>;
 confLH.chat_message_sinterval = <?php echo (int)(erConfigClassLhConfig::getInstance()->getSetting('chat','chat_message_sinterval')*1000) ?>;
-confLH.new_chat_sound_enabled = <?php echo erLhcoreClassModelUserSetting::getSetting('new_chat_sound',(int)(erConfigClassLhConfig::getInstance()->getSetting('chat','new_chat_sound_enabled'))) ?>;
-confLH.new_message_sound_admin_enabled = <?php echo erLhcoreClassModelUserSetting::getSetting('chat_message',(int)(erConfigClassLhConfig::getInstance()->getSetting('chat','new_message_sound_admin_enabled'))) ?>;
-confLH.new_message_sound_user_enabled = <?php echo erLhcoreClassModelUserSetting::getSetting('chat_message',(int)(erConfigClassLhConfig::getInstance()->getSetting('chat','new_message_sound_user_enabled'))) ?>;
+confLH.new_chat_sound_enabled = <?php echo (int)erLhcoreClassModelUserSetting::getSetting('new_chat_sound',(int)(erConfigClassLhConfig::getInstance()->getSetting('chat','new_chat_sound_enabled'))) ?>;
+confLH.new_message_sound_admin_enabled = <?php echo (int)erLhcoreClassModelUserSetting::getSetting('chat_message',(int)(erConfigClassLhConfig::getInstance()->getSetting('chat','new_message_sound_admin_enabled'))) ?>;
+confLH.new_message_sound_user_enabled = <?php echo (int)erLhcoreClassModelUserSetting::getSetting('chat_message',(int)(erConfigClassLhConfig::getInstance()->getSetting('chat','new_message_sound_user_enabled'))) ?>;
 confLH.transLation = {'new_chat':'<?php echo erTranslationClassLhTranslation::getInstance()->getTranslation('pagelayout/pagelayout','New chat request')?>'};
 confLH.csrf_token = '<?php echo erLhcoreClassUser::instance()->getCSFRToken()?>';
 </script>
